# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SBG Internal Portal API",
    "description": "REST API for the AWS Student Builder Group internal operations portal — members, tasks, submissions, link shortener, audit logs, and analytics. All endpoints (except auth) require a valid `sbg_session` HttpOnly cookie obtained via the OTP login flow.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment origin" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbg_session"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Shared secret configured via the CRON_SECRET env var. An `Authorization: Bearer <secret>` header is also accepted."
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["SBG_LEADER", "SECRETARY", "DIRECTOR", "MANAGER", "ASSOCIATE", "BUILDER"]
      },
      "Domain": {
        "type": "string",
        "enum": ["Technical", "Corporate", "Creatives", "General"]
      },
      "Subdomain": {
        "type": "string",
        "enum": ["Software Development", "AI & Machine Learning", "Cloud & DevOps", "Events & Operations", "Sponsorship & Finance", "HR & Admin", "PR & Marketing", "Digital Design", "Media Production"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "Member": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "clubId": { "type": "string" },
          "name": { "type": "string" },
          "regNo": { "type": "string" },
          "department": { "type": "string" },
          "section": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "officialEmail": { "type": "string", "format": "email" },
          "personalEmail": { "type": "string" },
          "phone": { "type": "string" },
          "whatsapp": { "type": "string" },
          "github": { "type": "string" },
          "linkedin": { "type": "string" },
          "instagram": { "type": "string" },
          "meetup": { "type": "string" },
          "builderId": { "type": "string" },
          "faName": { "type": "string" },
          "faEmail": { "type": "string" },
          "faPhone": { "type": "string" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" },
          "totalStars": { "type": "number" }
        }
      },
      "MemberCreateRequest": {
        "type": "object",
        "required": ["officialEmail"],
        "properties": {
          "officialEmail": { "type": "string", "format": "email" },
          "clubId": { "type": "string" },
          "name": { "type": "string" },
          "regNo": { "type": "string" },
          "department": { "type": "string" },
          "section": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain" },
          "subdomain": { "$ref": "#/components/schemas/Subdomain" },
          "personalEmail": { "type": "string" },
          "phone": { "type": "string" },
          "whatsapp": { "type": "string" },
          "github": { "type": "string" },
          "linkedin": { "type": "string" },
          "instagram": { "type": "string" },
          "meetup": { "type": "string" },
          "builderId": { "type": "string" },
          "faName": { "type": "string" },
          "faEmail": { "type": "string" },
          "faPhone": { "type": "string" }
        }
      },
      "MemberUpdateRequest": {
        "type": "object",
        "description": "Self-service callers may only set phone, github, linkedin, meetup, personalEmail. Presidium may additionally set name, role, domain, subdomain, department, builderId, isActive, clubId, regNo.",
        "properties": {
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain" },
          "subdomain": { "$ref": "#/components/schemas/Subdomain" },
          "department": { "type": "string" },
          "phone": { "type": "string" },
          "github": { "type": "string" },
          "linkedin": { "type": "string" },
          "meetup": { "type": "string" },
          "builderId": { "type": "string" },
          "personalEmail": { "type": "string" },
          "isActive": { "type": "boolean" },
          "clubId": { "type": "string" },
          "regNo": { "type": "string" }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "taskId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "assignmentType": { "type": "string", "enum": ["INDIVIDUAL", "DOMAIN", "SUBDOMAIN", "GENERAL"] },
          "assignedToId": { "type": "string", "nullable": true },
          "assignedToName": { "type": "string" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "createdBy": { "type": "string" },
          "createdByName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"], "description": "Lazily flipped OPEN->CLOSED on read once the deadline has passed — no scheduled job required." },
          "totalSubmissions": { "type": "integer" },
          "reminderSentAt": { "type": "string", "format": "date-time", "nullable": true, "description": "Set once a deadline-reminder email batch has been sent for this task; cleared if the deadline is changed." }
        }
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["LOW", "MEDIUM", "HIGH"]
      },
...
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time", "description": "Changing this clears reminderSentAt so a new reminder cycle can fire" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"] }
        }
      },
      "TaskDetail": {
        "type": "object",
        "properties": {
          "task": { "$ref": "#/components/schemas/Task" },
          "submissions": { "type": "array", "items": { "$ref": "#/components/schemas/Submission" } },
          "mySubmission": { "allOf": [{ "$ref": "#/components/schemas/Submission" }], "nullable": true },
          "canReview": { "type": "boolean" },
          "canSubmit": { "type": "boolean" },
          "canDelete": { "type": "boolean" },
          "canClose": { "type": "boolean" }
        }
      },
      "Submission": {
        "type": "object",
        "properties": {
          "submissionId": { "type": "string" },
          "taskId": { "type": "string" },
          "taskTitle": { "type": "string" },
          "memberId": { "type": "string" },
          "memberName": { "type": "string" },
          "content": { "type": "string" },
          "links": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "submittedAt": { "type": "string", "format": "date-time" },
          "reviewStatus": { "type": "string", "enum": ["PENDING", "APPROVED", "REJECTED"] },
          "reviewedBy": { "type": "string", "nullable": true },
          "reviewedByName": { "type": "string", "nullable": true },
          "reviewedAt": { "type": "string", "format": "date-time", "nullable": true },
          "reviewFeedback": { "type": "string", "nullable": true, "description": "Optional comment left by the reviewer on APPROVE/REJECT" },
          "ratingAwarded": { "type": "number", "nullable": true },
          "deadline": { "type": "string", "format": "date-time" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true }
        }
      },
...
      "ReviewRequest": {
        "type": "object",
        "required": ["submissionId", "action"],
        "properties": {
          "submissionId": { "type": "string" },
          "action": { "type": "string", "enum": ["APPROVE", "REJECT"] },
          "feedback": { "type": "string", "description": "Optional comment shown to the submitter" }
        }
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "memberName": { "type": "string" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "role": { "$ref": "#/components/schemas/Role" },
          "totalStars": { "type": "number" },
          "approvedCount": { "type": "integer" },
          "lateApprovedCount": { "type": "integer" },
          "rejectedCount": { "type": "integer" },
          "pendingCount": { "type": "integer" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Unauthorized" } } }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted to perform this action",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Forbidden" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/members": {
      "get": {
        "tags": ["Members"],
        "summary": "List members",
        "description": "Inactive members are hidden unless the caller is presidium and passes `all=true`.",
        "parameters": [
          { "name": "domain", "in": "query", "schema": { "$ref": "#/components/schemas/Domain" } },
          { "name": "subdomain", "in": "query", "schema": { "$ref": "#/components/schemas/Subdomain" } },
          { "name": "role", "in": "query", "schema": { "$ref": "#/components/schemas/Role" } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Matches name, officialEmail, or clubId (case-insensitive substring)" },
          { "name": "all", "in": "query", "schema": { "type": "boolean" }, "description": "Presidium-only: include inactive members" }
        ],
        "responses": {
          "200": { "description": "Member list", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Member" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Members"],
        "summary": "Create a member",
        "description": "Presidium only.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemberCreateRequest" } } } },
        "responses": {
          "200": { "description": "Member created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Member" } } } } } },
          "400": { "description": "officialEmail is required" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "description": "A member with that email already exists" }
        }
      }
    },
    "/members/{memberId}": {
      "get": {
        "tags": ["Members"],
        "summary": "Get a member by id (includes rating)",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Member detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "allOf": [{ "$ref": "#/components/schemas/Member" }, { "type": "object", "properties": { "rating": { "$ref": "#/components/schemas/Rating", "nullable": true } } }] } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Members"],
        "summary": "Update a member",
        "description": "Presidium may edit any field on any member. Non-presidium callers may only edit their own record, and only: phone, github, linkedin, meetup, personalEmail.",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemberUpdateRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "No valid fields to update" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Members"],
        "summary": "Deactivate a member (soft delete)",
        "description": "Presidium only. Sets isActive=false; does not remove the record.",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deactivated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/tasks/{taskId}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task detail with submissions and caller permissions",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/TaskDetail" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Task not found or not visible to caller" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task (title, description, deadline, status)",
        "description": "Only the task creator or presidium may edit. Reopening a CLOSED task requires reviewer privileges (presidium/director/manager).",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskUpdateRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid status value, or no valid fields to update" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task and cascade-delete its submissions",
        "description": "Only the task creator or presidium may delete.",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/tasks/{taskId}/review": {
      "post": {
        "tags": ["Submissions"],
        "summary": "Approve or reject a pending submission",
        "description": "Caller must have review rights over the submission's domain/subdomain scope (or be presidium). Approving awards a star rating based on submission timeliness vs. the task deadline.",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewRequest" } } } },
        "responses": {
          "200": { "description": "Reviewed", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "ratingAwarded": { "type": "number" } } } } } },
          "400": { "description": "Invalid request, or taskId/submission mismatch" },
          "403": { "description": "Not authorized to review this submission" },
          "404": { "description": "Submission or task not found" },
          "409": { "description": "Submission already reviewed" }
        }
      }
    },
...
  }
}